% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{AUC}
\alias{AUC}
\title{Compute the AUC for ROC values}
\usage{
AUC(x, y)
}
\arguments{
\item{x}{vector of \code{TPR} or \code{ROC} object}

\item{y}{vector of \code{FPR} or ignored if \code{x} inherits from class \code{ROC}}
}
\value{
numeric Area Under Curve
}
\description{
Compute the AUC for ROC values
}
