% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{FPR}
\alias{FPR}
\title{False Positive Rate = FP/N}
\usage{
FPR(x, threshold = seq(from = 0, by = 0.01, length = 100))
}
\arguments{
\item{x}{tibble with columns pres (0/1) and pred (0-1)}

\item{threshold, }{numeric values at or above are 'positive' predictions}
}
\value{
vector of FPR
}
\description{
False Positive Rate = FP/N
}
