% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot.ROC}
\alias{plot.ROC}
\title{Plot a ROC object}
\usage{
\method{plot}{ROC}(
  x,
  xlab = "False Positive Rate (Specificity)",
  ylab = "True Positive Rate (Sensitivity)",
  title = "ROC",
  use = c("base", "ggplot2", "ggplot")[2],
  ...
)
}
\arguments{
\item{x}{ROC class object - see \code{\link{ROC}}}

\item{xlab}{char, x axis label}

\item{ylab}{char, y axis label}

\item{title}{char, plot title}

\item{use}{char one of 'base', 'ggplot' or 'ggplot2'}

\item{...}{other arguments for \code{plot} or \code{ggplot2::labs}}
}
\value{
if \code{use} is 'ggplot' or 'ggplot2' a ggplot object otherwise NULL
}
\description{
Plot a ROC object
}
