% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presence-package.R
\docType{package}
\name{presence}
\alias{presence}
\alias{presence-package}
\title{Tools in R for working with presence only modeling.}
\description{
Tools in R for working with presence only modeling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bigelowlab/presence}
}

}
\keyword{internal}
