% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{qplot.ROC}
\alias{qplot.ROC}
\title{Plot a ROC using ggplot2}
\usage{
qplot.ROC(
  x,
  xlab = "False Positive Rate (Specificity)",
  ylab = "True Positive Rate (Sensitivity)",
  title = "ROC",
  ...
)
}
\arguments{
\item{x}{ROC class object - see \code{\link{ROC}}}

\item{xlab}{char, x axis label}

\item{ylab}{char, y axis label}

\item{title}{char, plot title}

\item{...}{other arguments for  \code{\link[ggplot2]{labs}}}
}
\value{
ggplot2 object
}
\description{
Plot a ROC using ggplot2
}
