% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{ROC}
\alias{ROC}
\title{Compute Receiver Operator Curve (ROC) space values}
\usage{
ROC(x, threshold = seq(from = 0, by = 0.01, length = 100))
}
\arguments{
\item{x}{tibble with columns pres (0/1) and pred (0-1)}

\item{threshold, }{numeric values at or above are 'positive' predictions}
}
\value{
data frame of class ROC with variables \code{tpr} and \code{fpr}
}
\description{
Compute Receiver Operator Curve (ROC) space values
}
